+++ sen/parser_test.go
@@ -3,11 +3,13 @@
 package sen_test
 
 import (
+	"encoding/json"
 	"fmt"
 	"strings"
 	"testing"
 	"testing/iotest"
 
+	"github.com/ohler55/ojg"
 	"github.com/ohler55/ojg/sen"
 	"github.com/ohler55/ojg/tt"
 )
@@ -497,4 +499,18 @@ func TestParserTokenFunc(t *testing.T) {
 	src = strings.Repeat(" ", 4090) + "fun(abc)"
 	v = sen.MustParseReader(strings.NewReader(src))
 	tt.Equal(t, "abc", v)
+}
+
+func TestParserNumConv(t *testing.T) {
+	v := sen.MustParse([]byte("0.1234567890123456789"))
+	tt.Equal(t, json.Number("0.1234567890123456789"), v)
+
+	v = sen.MustParse([]byte("0.1234567890123456789"), ojg.NumConvFloat64)
+	tt.Equal(t, 0.123456789012345678, v)
+
+	v = sen.MustParse([]byte("0.1234567890123456789"), ojg.NumConvString)
+	tt.Equal(t, "0.1234567890123456789", v)
+
+	v = sen.MustParseReader(strings.NewReader("0.1234567890123456789"), ojg.NumConvFloat64)
+	tt.Equal(t, 0.123456789012345678, v)
 }