+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 The structure and content of this file follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.19.0] - 2023-07-05
+### Added
+- Added NumConvMethod to convert json.Number to either float64 or a
+  string on parse and recompose.
+
 ## [1.18.7] - 2023-06-02
 ### Fixed
 - Fixed alt.Diff to not skip non-matching map entries.